gt5-common: Hexedit sec-ril libs to allow audioserver connections
This is an old blob from a system which only had mediaserver. In
N audioserver has been created. So you need to search for the string:

  /system/bin/mediaserver

You need to replace 'mediaserver' with 'audioserver'. Right before
this string you will find the string 'media' which is group the
process is running as, replace that with 'audio'.

Change-Id: Ib40073237c7923a48ca08be62f2c130359defa8c